+++ src/routes/companies/company.industry.ts
